+++ packages/ui/src/components/widgets/SwapWidgetRenderer.tsx
@@ -22,7 +22,7 @@ import {
   isHighRelayerServiceFeeUsd,
   parseFees
 } from '../../utils/quote.js'
-import { usePrice } from '@reservoir0x/relay-kit-hooks'
+import { usePrice, useQuote } from '@reservoir0x/relay-kit-hooks'
 import { EventNames } from '../../constants/events.js'
 import { ProviderOptionsContext } from '../../providers/RelayKitProvider.js'
 import type { DebouncedState } from 'usehooks-ts'
@@ -93,7 +93,9 @@ export type ChildrenProps = {
   setAmountOutputValue: (value: string) => void
   debouncedAmountOutputControls: DebouncedState<(value: string) => void>
   toBalance?: bigint
+  toBalancePending?: boolean
   fromBalance?: bigint
+  fromBalancePending?: boolean
   isFetchingPrice: boolean
   isLoadingToBalance: boolean
   isLoadingFromBalance: boolean
@@ -228,7 +230,8 @@ const SwapWidgetRenderer: FC<SwapWidgetRendererProps> = ({
     queryKey: fromBalanceQueryKey,
     isLoading: isLoadingFromBalance,
     isError: fromBalanceErrorFetching,
-    isDuneBalance: fromBalanceIsDune
+    isDuneBalance: fromBalanceIsDune,
+    hasPendingBalance: fromBalancePending
   } = useCurrencyBalance({
     chain: fromChain,
     address: address,
@@ -240,7 +243,8 @@ const SwapWidgetRenderer: FC<SwapWidgetRendererProps> = ({
     value: toBalance,
     queryKey: toBalanceQueryKey,
     isLoading: isLoadingToBalance,
-    isDuneBalance: toBalanceIsDune
+    isDuneBalance: toBalanceIsDune,
+    hasPendingBalance: toBalancePending
   } = useCurrencyBalance({
     chain: toChain,
     address: recipient,
@@ -343,12 +347,7 @@ const SwapWidgetRenderer: FC<SwapWidgetRendererProps> = ({
       ? true
       : false
 
-  const {
-    data: price,
-    isLoading: isFetchingPrice,
-    error
-  } = usePrice(
-    relayClient ? relayClient : undefined,
+  const quoteParameters =
     fromToken && toToken
       ? {
           user: fromAddressWithFallback,
@@ -372,33 +371,60 @@ const SwapWidgetRenderer: FC<SwapWidgetRendererProps> = ({
           referrer: relayClient?.source ?? undefined,
           useExternalLiquidity
         }
-      : undefined,
-    ({ details }) => {
-      onAnalyticEvent?.(EventNames.SWAP_EXECUTE_QUOTE_RECEIVED, {
-        wallet_connector: connector?.name,
-        amount_in: details?.currencyIn?.amountFormatted,
-        currency_in: details?.currencyIn?.currency?.symbol,
-        chain_id_in: details?.currencyIn?.currency?.chainId,
-        amount_out: details?.currencyOut?.amountFormatted,
-        currency_out: details?.currencyOut?.currency?.symbol,
-        chain_id_out: details?.currencyOut?.currency?.chainId,
-        is_canonical: useExternalLiquidity
-      })
-    },
+      : undefined
+
+  const onQuoteReceived: Parameters<typeof usePrice>['2'] = ({ details }) => {
+    onAnalyticEvent?.(EventNames.SWAP_EXECUTE_QUOTE_RECEIVED, {
+      wallet_connector: connector?.name,
+      amount_in: details?.currencyIn?.amountFormatted,
+      currency_in: details?.currencyIn?.currency?.symbol,
+      chain_id_in: details?.currencyIn?.currency?.chainId,
+      amount_out: details?.currencyOut?.amountFormatted,
+      currency_out: details?.currencyOut?.currency?.symbol,
+      chain_id_out: details?.currencyOut?.currency?.chainId,
+      is_canonical: useExternalLiquidity
+    })
+  }
+
+  const quoteFetchingEnabled = Boolean(
+    relayClient &&
+      ((tradeType === 'EXACT_INPUT' &&
+        debouncedInputAmountValue &&
+        debouncedInputAmountValue.length > 0 &&
+        Number(debouncedInputAmountValue) !== 0) ||
+        (tradeType === 'EXPECTED_OUTPUT' &&
+          debouncedOutputAmountValue &&
+          debouncedOutputAmountValue.length > 0 &&
+          Number(debouncedOutputAmountValue) !== 0)) &&
+      fromToken !== undefined &&
+      toToken !== undefined
+  )
+
+  const {
+    data: _priceData,
+    isLoading: _isFetchingPrice,
+    error: priceError
+  } = usePrice(
+    relayClient ? relayClient : undefined,
+    quoteParameters,
+    onQuoteReceived,
     {
-      enabled: Boolean(
-        relayClient &&
-          ((tradeType === 'EXACT_INPUT' &&
-            debouncedInputAmountValue &&
-            debouncedInputAmountValue.length > 0 &&
-            Number(debouncedInputAmountValue) !== 0) ||
-            (tradeType === 'EXPECTED_OUTPUT' &&
-              debouncedOutputAmountValue &&
-              debouncedOutputAmountValue.length > 0 &&
-              Number(debouncedOutputAmountValue) !== 0)) &&
-          fromToken !== undefined &&
-          toToken !== undefined
-      ),
+      enabled: quoteFetchingEnabled
+    }
+  )
+
+  const {
+    data: _quoteData,
+    error: quoteError,
+    isLoading: isFetchingQuote
+  } = useQuote(
+    relayClient ? relayClient : undefined,
+    wallet,
+    quoteParameters,
+    undefined,
+    onQuoteReceived,
+    {
+      enabled: quoteFetchingEnabled,
       refetchInterval:
         !transactionModalOpen &&
         debouncedInputAmountValue === amountInputValue &&
@@ -407,6 +433,11 @@ const SwapWidgetRenderer: FC<SwapWidgetRendererProps> = ({
           : undefined
     }
   )
+
+  //Here we fetch the price data and quote data in parallel and then merge into one data model
+  const isFetchingPrice = isFetchingQuote ?? _isFetchingPrice
+  const error = _quoteData || isFetchingQuote ? null : quoteError ?? priceError
+  const price = error ? undefined : _quoteData ?? _priceData
 
   useDisconnected(address, () => {
     setCustomToAddress(undefined)
@@ -484,9 +515,10 @@ const SwapWidgetRenderer: FC<SwapWidgetRendererProps> = ({
   const isInsufficientLiquidityError = fetchQuoteErrorMessage?.includes(
     'No quotes available'
   )
-  const isCapacityExceededError = fetchQuoteDataErrorMessage?.includes(
-    'Amount is higher than the available liquidity'
-  )
+  const isCapacityExceededError =
+    fetchQuoteDataErrorMessage?.includes(
+      'Amount is higher than the available liquidity'
+    ) || fetchQuoteDataErrorMessage?.includes('Insufficient relayer liquidity')
   const isCouldNotExecuteError =
     fetchQuoteDataErrorMessage?.includes('Could not execute')
   const highRelayerServiceFee = isHighRelayerServiceFeeUsd(price)
@@ -623,10 +655,12 @@ const SwapWidgetRenderer: FC<SwapWidgetRendererProps> = ({
         debouncedAmountOutputControls,
         setAmountOutputValue,
         toBalance,
+        toBalancePending,
         isLoadingToBalance,
         isFetchingPrice,
         isLoadingFromBalance,
         fromBalance,
+        fromBalancePending,
         highRelayerServiceFee,
         relayerFeeProportion,
         hasInsufficientBalance,