+++ packages/ui/src/hooks/useDuneBalances.ts
@@ -6,7 +6,8 @@ import {
   type DefaultError,
   type QueryKey
 } from '@tanstack/react-query'
-import { solana, solanaAddressRegex } from '../utils/solana.js'
+import { isSolanaAddress, solana } from '../utils/solana.js'
+import { isBitcoinAddress } from '../utils/bitcoin.js'
 
 export type DuneBalanceResponse = {
   request_time: string
@@ -22,7 +23,7 @@ export type DuneBalanceResponse = {
     price_usd?: number
     value_usd?: number
   }>
-}
+} | null
 
 type QueryType = typeof useQuery<
   DuneBalanceResponse,
@@ -35,7 +36,8 @@ type QueryOptions = Parameters<QueryType>['0']
 export default (address?: string, queryOptions?: Partial<QueryOptions>) => {
   const providerOptions = useContext(ProviderOptionsContext)
   const queryKey = ['useDuneBalances', address]
-  const isSvmAddress = address && solanaAddressRegex.test(address)
+  const isSvmAddress = isSolanaAddress(address ?? '')
+  const isBvmAddress = isBitcoinAddress(address ?? '')
 
   const response = (useQuery as QueryType)({
     queryKey: ['useDuneBalances', address],
@@ -43,6 +45,10 @@ export default (address?: string, queryOptions?: Partial<QueryOptions>) => {
       let url = `https://api.dune.com/api/beta/balance/${address?.toLowerCase()}?chain_ids=all&exclude_spam_tokens=true`
       if (isSvmAddress) {
         url = `https://api.dune.com/api/beta/balance/solana/${address}?chain_ids=all&exclude_spam_tokens=true`
+      }
+
+      if (isBvmAddress) {
+        return null
       }
 
       return fetch(url, {
@@ -54,7 +60,7 @@ export default (address?: string, queryOptions?: Partial<QueryOptions>) => {
         .then((response) => {
           if (response.balances) {
             const balances =
-              response.balances as DuneBalanceResponse['balances']
+              response.balances as NonNullable<DuneBalanceResponse>['balances']
             if (balances) {
               balances
                 .filter((balance) => {
@@ -88,17 +94,21 @@ export default (address?: string, queryOptions?: Partial<QueryOptions>) => {
           return response
         })
     },
-    enabled: address !== undefined && providerOptions.duneApiKey !== undefined,
-    ...queryOptions
+    ...queryOptions,
+    enabled:
+      address !== undefined &&
+      providerOptions.duneApiKey !== undefined &&
+      queryOptions?.enabled &&
+      !isBvmAddress
   })
 
-  response.data?.balances?.forEach((balance) => {
+  response?.data?.balances?.forEach((balance) => {
     if (!balance.chain_id && balance.chain === 'solana') {
       balance.chain_id = solana.id
     }
   })
 
-  const balanceMap = response.data?.balances?.reduce((balanceMap, balance) => {
+  const balanceMap = response?.data?.balances?.reduce((balanceMap, balance) => {
     if (balance.address === 'native') {
       balance.address =
         balance.chain === 'solana'
@@ -112,7 +122,7 @@ export default (address?: string, queryOptions?: Partial<QueryOptions>) => {
 
     balanceMap[`${chainId}:${balance.address}`] = balance
     return balanceMap
-  }, {} as Record<string, DuneBalanceResponse['balances'][0]>)
+  }, {} as Record<string, NonNullable<DuneBalanceResponse>['balances'][0]>)
 
   return { ...response, balanceMap, queryKey } as ReturnType<QueryType> & {
     balanceMap: typeof balanceMap